requirements.txt
APScheduler==3.6.3
astroid==2.9.0
attrs==21.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.4
click==8.0.1
dataclasses==0.8
decorator==4.4.2
flake8==4.0.1
greenlet==1.1.1
idna==3.2
importlab==0.6.1
importlib-metadata==4.2.0
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.6.0
libcst==0.3.23
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.5.1
ninja==1.10.2.3
ox-secrets==0.1.0
packaging==21.0
pkg-resources==0.0.0
platformdirs==2.4.0
pluggy==0.13.1
py==1.10.0
pybind11==2.8.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.1
pyparsing==2.4.7
pytest==6.2.4
python-telegram-bot==13.8.1
pytype==2021.11.24
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
six==1.16.0
tabulate==0.8.9
toml==0.10.2
tornado==6.1
typed-ast==1.5.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
wrapt==1.13.3
zipp==3.6.0
